Allow float values for cluster distance, re #7120

+++ arches/app/views/api.py
@@ -404,7 +404,7 @@ class MVT(APIBase):
             with connection.cursor() as cursor:
                 if int(zoom) <= int(config["clusterMaxZoom"]):
                     arc = self.EARTHCIRCUM / ((1 << int(zoom)) * self.PIXELSPERTILE)
-                    distance = arc * int(config["clusterDistance"])
+                    distance = arc * float(config["clusterDistance"])
                     min_points = int(config["clusterMinPoints"])
                     cursor.execute(
                         """WITH clusters(tileid, resourceinstanceid, nodeid, geom, cid)
